requirements.txt
pypinfo
PyGithub
cookieninja
grpcio-status
mypy
pre-commit
pycloc
pytest
types-requests
